fix: split snapshot capture tests and type surface
